api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "django-ninja>=1.5.3",
    "django-stubs>=5.2.9",
    "django-types>=0.23.0",
]
